#31 small improvements

+++ Storage/Repositories/Models/MeetingSeatUpdate.cs
@@ -6,7 +6,7 @@
 
         public Guid EventID { get; set; }
 
-        public int SequenceNumber { get; set; }
+        public long SequenceNumber { get; set; }
 
         public DateTime Timestamp { get; set; }
     }
